+++ src/main/java/hr/avrbanac/inanis/internals/GameState.java
@@ -1,0 +1,7 @@
+package hr.avrbanac.inanis.internals;
+
+public enum GameState {
+    INIT,
+    STARTED,
+    GAME_OVER
+}